feat: Adds location module and integrates with DogList
Implements a new location module for both iOS and Android using TurboModules.

The module allows the app to request location permissions and retrieve the current location of the user.

Integrates the location module with the DogList component to display dogs sorted by distance from the user. The DogList component also now includes a refresh button and an update location button.
Also, adds missing font files to the Info.plist.