feat: Add TurboModule for Location request (#1)
* feat: Adds location module and integrates with DogList

Implements a new location module for both iOS and Android using TurboModules.

The module allows the app to request location permissions and retrieve the current location of the user.

Integrates the location module with the DogList component to display dogs sorted by distance from the user. The DogList component also now includes a refresh button and an update location button.
Also, adds missing font files to the Info.plist.

* Displays dog distance and nearby badge

Integrates the `react-native-location` module to display the distance of each dog from the user's location.

Enhances the user experience by adding a "nearby" badge to dogs that are less than 1 km away from the user.

* chore: Refactors location permission check and UI tweaks

Consolidates location permission checks to use the location manager instance for consistency.

Improves UI by reducing the font size of the button text in the DogList component, and removes unused imports from the DogCard component.